+++ addon/background.js
@@ -1,4 +1,6 @@
 "use strict";
+let sfHost;
+
 chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
   // Perform cookie operations in the background page, because not all foreground pages have access to the cookie API.
   // Firefox does not support incognito split mode, so we use sender.tab.cookieStoreId to select the right cookie store.
@@ -32,6 +34,7 @@ chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
     return true; // Tell Chrome that we want to call sendResponse asynchronously.
   }
   if (request.message == "getSession") {
+    sfHost = request.sfHost;
     chrome.cookies.get({url: "https://" + request.sfHost, name: "sid", storeId: sender.tab.cookieStoreId}, sessionCookie => {
       if (!sessionCookie) {
         sendResponse(null);
@@ -44,6 +47,19 @@ chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
   }
   return false;
 });
+
+chrome.commands.onCommand.addListener((command) => {
+  if (command !== "open-popup"){
+    chrome.tabs.create({
+      url: `chrome-extension://${chrome.i18n.getMessage("@@extension_id")}/${command}.html?host=${sfHost}`
+    });
+  } else {
+    chrome.runtime.sendMessage({
+      msg: "shortcut_pressed", command, sfHost
+    });
+  }
+});
+
 chrome.runtime.onInstalled.addListener(({reason}) => {
   if (reason === "install") {
     chrome.tabs.create({